API Reference

Public

GenomicFeatures.STRAND_BOTH — Constant

Both strand ('.')

GenomicFeatures.STRAND_NA — Constant

Unknown strand ('?')

GenomicFeatures.STRAND_NEG — Constant

Negative strand ('-')

GenomicFeatures.STRAND_POS — Constant

Positive strand ('+')

GenomicFeatures.Interval — Type

struct Interval{T} <: IntervalTrees.AbstractInterval{Int64}

The first three fields (seqname, first, and last) are mandatory arguments when constructing the Interval object.

Fields

seqname::String: the sequence name associated with the interval.

first::Int64: the leftmost position.

last::Int64: the rightmost position.

strand::Strand: the strand.

metadata::T

GenomicFeatures.Interval — Method

Interval{T}(data)

The returned data is converted to Interval{T} if there is an implemented Base.convert function for the type of data. This method provides a useful hook for converting custom types to Interval{T}.

GenomicFeatures.IntervalCollection — Type

An IntervalCollection is an efficiently stored and indexed set of annotated genomic intervals.

GenomicFeatures.IntervalCollection — Type

IntervalCollection(data, sort::Bool=false)

Constructor that guesses metadatatype, and offers conversion through collection.

GenomicFeatures.IntervalCollection — Method

IntervalCollection(intervals::AbstractVector{Interval{T}}, sort::Bool=false) where T

Shorthand constructor.

GenomicFeatures.IntervalCollection — Method

IntervalCollection{T}(data, sort::Bool=false) where T

Constructor that offers conversion through collection.

GenomicFeatures.Strand — Type

Outer constructors

Strand(strand::Char)

Strand(strand::UInt8)

Strand can take four kinds of values listed in the next table:

Symbol	Constant	Meaning

'?'	STRAND_NA	strand is unknown or inapplicable

'+'	STRAND_POS	positive strand

'-'	STRAND_NEG	negative strand

'.'	STRAND_BOTH	non-strand-specific feature

GenomicFeatures.Strand — Method

Strand(strand::Char)

GenomicFeatures.Strand — Method

Strand(strand::UInt8)

BioGenerics.isoverlapping — Method

Return true if interval a overlaps interval b, with no consideration to strand

BioGenerics.leftposition — Method

leftposition(i::Interval)

Return the leftmost position of i.

BioGenerics.rightposition — Method

rightposition(i::Interval)

Return the rightmost position of i.

GenomicFeatures.coverage — Function

coverage(intervals)

Compute the coverage of a collection of intervals and return an IntervalCollection that contains run-length encoded coverage data.

For example, given intervals like:

[------]     [------------]
   [---------------]

This function would return a new set of disjoint intervals with annotated coverage like:

[1][-2-][-1-][--2--][--1--]

Example

julia> intervals = [
           Interval("chr1", 1, 8),
           Interval("chr1", 4, 20),
           Interval("chr1", 14, 27)];

julia> coverage(intervals)
IntervalCollection{UInt32} with 5 intervals:
  chr1:1-3  .  1
  chr1:4-8  .  2
  chr1:9-13  .  1
  chr1:14-20  .  2
  chr1:21-27  .  1

GenomicFeatures.eachoverlap — Function

eachoverlap(intervals_a, intervals_b, [seqname_isless=Base.isless])

Create an iterator of overlapping intervals between intervals_a and intervals_b.

This function assumes elements of intervals_a and intervals_b are sorted by its sequence name and left position. If the element type is not a subtype of GenomicFeatures.Interval, elements are converted to Interval objects.

The third optional argument is a function that defines the order of sequence names. The default function is Base.isless, which is the lexicographical order.

Internal

Base.findfirst — Method

Find a the first interval with matching start and end points.

Returns that interval, or 'nothing' if no interval was found.

GenomicFeatures.isordered — Method

Check if two intervals are well ordered.

Intervals are considered well ordered if a.seqname <= b.seqnamend and a.first <= b.first.

GenomicFeatures.precedes — Method

Return true if interval a entirely precedes b.